Add ORMLite integration and tests
ORMLite enables much easier access to fetching and creating queries over
the JDBC database. This requires additional entities for many-to-many
relationships: recipes can have multiple tags, and a tag can be
associated with multiple recipes.

Access to the ORM is done through DAOs (Data Access Objects), which are
instantiated by the Database class.